integration to main

+++ lab2/software/ps_matmul/src/user_uart.h
@@ -15,7 +15,7 @@ int UART_Init(XUartPs *Uart_Ps, u16 DeviceId);
 int UART_Init(XUartPs *Uart_Ps, UINTPTR BaseAddress);
 #endif
 
-u32 UART_RxToBuffer(XUartPs *Uart_Ps, u8 *BufferPtr, u32 NumBytes);
+u32 UART_RxToBuffer(XUartPs *Uart_Ps, u32 *BufferPtr, u32 NumBytes);
 u32 UART_TxFromBuffer(XUartPs *Uart_Ps, u8 *BufferPtr, u32 NumBytes);
 
 #endif /* USER_UART_H */
